fix routers and made local storage

+++ backend/todo.route.js
@@ -1,17 +1,40 @@
 const express = require('express');
 const todoRoutes = express.Router();
 let Todo = require('./todo.model');
-const verifyToken = require('./auth/verifyToken');
+// const verifyToken = require('./auth/verifyToken');
 let jwt = require('jsonwebtoken');
 let config = require('../backend/config');
 
+//return user id on success or empty string
+function checkTokenAccess(token) {
+    let userID = '';
+    jwt.verify(
+      token,
+      config.secret,
+      (err, decoded) => {
+          if (err)
+              return null;
+          userID = decoded.id;
+      });
+    return userID;
+}
+
 todoRoutes.post(
-  '/add',
+  '/add/:id',
   (req, res) => {
+      let token = req.params.id;
+      let userID = checkTokenAccess(token);
+      if ('' === userID) {
+          res.status(404);
+          return;
+      }
       let todo = new Todo(req.body);
+      todo.user_id = userID;
+      todo.create_data = new Date().toLocaleDateString();
+      todo.completed = false;
       todo.save()
-        .then(todo => {
-            res.status(200).json({'todo': 'todo added successfully'});
+        .then((todo) => {
+            res.status(200).json(todo);
             console.log(todo);
         })
         .catch(err => {
@@ -21,48 +44,47 @@ todoRoutes.post(
   });
 
 todoRoutes.get(
-  '/:id',
+  '/',
   (req, res) => {
-      let token = req.params.id;
-      let userID = '';
-      jwt.verify(
-        token,
-        config.secret,
-        (err, decoded) => {
-            if (err)
-                return res.status(500).send({
-                    auth: false,
-                    message: 'Failed to authenticate token.'
-                });
+      console.log(req.get('authorization'));
 
-            // if everything good, save to request for use in other routes
-            userID = decoded.id;
-        });
-      Todo.findById(
-        userID,
-        (err, todos) => {
-            if (err) {
-                console.log(err);
-            } else {
-                res.json(todos);
-            }
-        });
+      // check for basic auth header
+      if (!req.headers.authorization || req.headers.authorization.indexOf('Basic ') === -1) {
+          return res.status(401).json({ err: 'Missing Authorization Header' });
+      }
+
+      // verify auth credentials
+      const token =  req.headers.authorization.split(' ')[1];
+      console.log("the token is ", token);
+      let userID = checkTokenAccess(token);
+      if ('' === userID) {
+          res.status(401).json({err: 'user id not found'});
+      } else {
+          Todo.find(
+            {user_id: userID},
+            (err, todos) => {
+                if (err) {
+                    res.status(404).json({err: 'TODO not found'});
+                    console.log(err);
+                } else if (null === todos) {
+                    res.status(200);
+                } else {
+                    res.status(200).send(todos.toString());
+                    // res.json(todos);
+                }
+            });
+      }
   });
-
-// todoRoutes.get(
-//   '/:id',
-//   (req, res) => {
-//       let id = req.params.id;
-//       Todo.findById(
-//         id,
-//         (err, todo) => {
-//             res.json(todo);
-//         });
-//   });
 
 todoRoutes.put(
   '/:id',
   (req, res) => {
+      let token = req.params.id;
+      let userID = checkTokenAccess(token);
+      if ('' === userID) {
+          res.status(404);
+          return;
+      }
       Todo.findById(
         req.params.id,
         (err, todo) => {
@@ -73,7 +95,6 @@ todoRoutes.put(
                 todo.user_id = req.body.user_id;
                 todo.completed = req.body.completed;
                 todo.create_data = req.body.create_data;
-
                 todo.save()
                   .then(todo => {
                       res.json('Todos updated!');
@@ -88,6 +109,12 @@ todoRoutes.put(
 todoRoutes.delete(
   '/:id',
   (req, res) => {
+      let token = req.params.id;
+      let userID = checkTokenAccess(token);
+      if ('' === userID) {
+          res.status(404);
+          return;
+      }
       console.log('Hei');
       Todo.findByIdAndRemove(
         req.params.id,
@@ -95,7 +122,7 @@ todoRoutes.delete(
             if (err) {
                 res.status(404).send("data is not found");
                 console.log("error '/delete/:id': ", todo);
-            }else{
+            } else {
                 res.json('Successfully removed');
             }
         })
